requirements/local.txt
-r base.txt

django_debug_toolbar==3.8.1
pre-commit==2.21.0
psycopg2-binary==2.9.5